LayoutTests/fast/events/ios/key-command-delete-to-end-of-paragraph-expected.txt

Tests that pressing Control + k in a content-editable field deletes to the end of the paragraph.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


To run this test by hand, place the text insertion point before the 'h' in the text below and press Control + k.

PASS document.getElementById("test").textContent is "a"
PASS successfullyParsed is true

TEST COMPLETE
